+++ crates/block-producer/src/lib.rs
@@ -6,9 +6,13 @@ pub mod test_utils;
 mod batch_builder;
 mod block_builder;
 mod domain;
-mod errors;
 mod mempool;
 pub mod store;
+
+#[cfg(feature = "testing")]
+pub mod errors;
+#[cfg(not(feature = "testing"))]
+mod errors;
 
 pub mod server;