+++ crates/utils/src/errors.rs
@@ -25,3 +25,18 @@ pub enum ApiError {
     #[error("parsing store url failed: {0}")]
     InvalidStoreUrl(String),
 }
+
+pub trait ErrorReport: std::error::Error {
+    fn as_report(&self) -> String {
+        use std::fmt::Write;
+        let mut report = self.to_string();
+
+        // SAFETY: write! is suggested by clippy, and is trivially safe usage.
+        std::iter::successors(self.source(), |child| child.source())
+            .for_each(|source| write!(report, "\nCaused by: {source}").unwrap());
+
+        report
+    }
+}
+
+impl<T: std::error::Error> ErrorReport for T {}